+++ presentation/src/main/java/com/knocklock/presentation/ui/setting/menu/Menu.kt
@@ -15,30 +15,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.knocklock.presentation.ui.setting.ActivatePassword
+import com.knocklock.presentation.ui.setting.ChangePassword
+import com.knocklock.presentation.ui.setting.Credit
+import com.knocklock.presentation.ui.setting.Menu
 
-interface Menu {
-    val title: String
-    val isNeedSwitch: Boolean
-    val route: String
-}
-
-object ActivatePassword : Menu {
-    override val title = "비밀번호 활성화"
-    override val isNeedSwitch = true
-    override val route = "activate_password"
-}
-
-object ChangePassword : Menu {
-    override val title = "비밀번호 변경"
-    override val isNeedSwitch = false
-    override val route = "change_password"
-}
-
-object Credit : Menu {
-    override val title = "크레딧"
-    override val isNeedSwitch = false
-    override val route = "credit"
-}
 
 @Composable
 fun MenuItem(
